testy

+++ src/main/java/org/tobiaszpietryga/payment/doman/Customer.java
@@ -26,7 +26,7 @@ public class Customer {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private String name;
-	private BigDecimal amountAvailable;
-	private BigDecimal amountReserved;
+	private Integer amountAvailable;
+	private Integer amountReserved;
 
 }
